pyproject.toml
[project]
name = "agentic-program-technical-assets"
version = "0.1.0"
description = "Technical assets for the Agentic Program. This is a top level pyproject to make development easier"
readme = "README.md"
requires-python = "==3.12.8"
# Keep only truly core dependencies here that every part of your project needs
dependencies = [
    "boto3>=1.37.13",
    "pydantic>=2.10.6",
    "fastapi>=0.115.11",
    "jupyter>=1.1.1",
    "langgraph>=0.3.11",
    "mcp>=1.5.0",
    "pytest>=8.3.5",
    "uvicorn>=0.34.0",
    "chromadb>=0.6.3",
    "llama-index>=0.12.25",
    "scikit-learn>=1.6.1",
    "langchain-aws>=0.2.18",
    "langchain-core>=0.3.50",
    "nest-asyncio>=1.6.0",
    "langchain-mcp-adapters>=0.0.7",
    "pydantic-ai-slim[bedrock,mcp,tavily,openai]>=0.2.14",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.5.4",
    "ipython>=9.0.2",
    "pydantic-graph>=0.2.4",
    "opentelemetry-api>=1.31.1",
    "opentelemetry-sdk>=1.31.1",
    "opentelemetry-exporter-otlp>=1.31.1",
    "opentelemetry-instrumentation-bedrock>=0.39.0",
    "opentelemetry-instrumentation-langchain>=0.39.0",
    "alembic>=1.15.2",
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.40",
    "pgvector>=0.4.0",
    "redis>=5.2.1",
    "ruff>=0.11.5",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.2",
    "litellm>=1.66.2",
    "pydantic-evals>=0.2.4",
    "ragas>=0.2.14",
    "strands-agents[litellm,openai]>=1.0.1",
    "strands-agents-tools>=0.1.9",
    "psycopg2-binary>=2.9.10",
    "langchain-litellm>=0.2.2",
]

[tool.uv.workspace]
members = ["labs/module4/notebooks/mcp", "labs/module4/mcp_server/weather_server", "labs/module4/notebooks/common/mcp_server/weather_server"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["agentic_platform"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
    "httpx>=0.27.0",
]
